+++ client.go
@@ -31,6 +31,7 @@ type Client struct {
 	userAgent  string
 
 	IPs             *IPService
+	LoadBalancers   *LoadBalancerService
 	Servers         *ServerService
 	Networks        *NetworkService
 	NetworkAdapters *NetworkAdapterService
@@ -50,6 +51,7 @@ func NewClient(project, apiKey, userAgent string) *Client {
 	}
 
 	c.IPs = &IPService{client: c}
+	c.LoadBalancers = &LoadBalancerService{client: c}
 	c.Servers = &ServerService{client: c}
 	c.Networks = &NetworkService{client: c}
 	c.NetworkAdapters = &NetworkAdapterService{client: c}